refactor: move banner engine into the client and drop the vendored builder copy
